backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.1.0
opencv-python==4.8.1.78
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.2
python-multipart==0.0.6
